requirements.txt
streamlit
scikit-learn
pillow
joblib
streamlit-drawable-canvas
numpy
